// next-starter-kit/care-pulse/src/stores/gold-price/types.ts
export interface Data {
  timestamp: number;
  previousClosePrice: number;
  openingPrice: number;
  lowPrice: number;
  highPrice: number;
  openTime: number;
  currentPrice: number;
  priceChange: number;
  priceChangePercentage: number;
  askPrice: number;
  bidPrice: number;
  priceGram24k: number;
  priceGram23k: number;
  priceGram22k: number;
  priceGram21k: number;
  key: string;
}